move qubitunitary to new module
