+++ pennylane/estimator/__init__.py
@@ -45,6 +45,7 @@ from .ops.qubit import (
     MultiRZ,
     PauliRot,
     SingleExcitation,
+    QubitUnitary,
 )
 
 from .ops.op_math import (
@@ -79,7 +80,6 @@ from .ops.templates import (
     Select,
     QROM,
     SelectPauliRot,
-    QubitUnitary,
     ControlledSequence,
     QPE,
     IterativeQPE,
