+++ pennylane/estimator/resource_config.py
@@ -18,7 +18,8 @@ from collections.abc import Callable
 from enum import StrEnum
 from typing import TYPE_CHECKING
 
-from pennylane.estimator.ops.templates import QubitUnitary, SelectPauliRot
+from pennylane.estimator.ops.qubit import QubitUnitary
+from pennylane.estimator.ops.templates import SelectPauliRot
 
 if TYPE_CHECKING:
     from pennylane.estimator.resource_operator import ResourceOperator
