+++ tests/estimator/templates/test_estimator_subroutines.py
@@ -1644,132 +1644,6 @@ class TestResourceQROM:
         )
 
 
-class TestResourceQubitUnitary:
-    """Test the ResourceQubitUnitary template"""
-
-    @pytest.mark.parametrize("precision", (None, 1e-3, 1e-5))
-    @pytest.mark.parametrize("num_wires", (1, 2, 3, 4, 5, 6))
-    def test_resource_params(self, num_wires, precision):
-        """Test that the resource params are correct."""
-        op = qre.QubitUnitary(num_wires, precision) if precision else qre.QubitUnitary(num_wires)
-        assert op.resource_params == {"num_wires": num_wires, "precision": precision}
-
-    @pytest.mark.parametrize("precision", (None, 1e-3, 1e-5))
-    @pytest.mark.parametrize("num_wires", (1, 2, 3, 4, 5, 6))
-    def test_resource_rep(self, num_wires, precision):
-        """Test that the compressed representation is correct."""
-        expected = qre.CompressedResourceOp(
-            qre.QubitUnitary, num_wires, {"num_wires": num_wires, "precision": precision}
-        )
-        assert qre.QubitUnitary.resource_rep(num_wires=num_wires, precision=precision) == expected
-
-    @pytest.mark.parametrize(
-        "num_wires, precision, expected_res",
-        (
-            (
-                1,
-                None,
-                [
-                    GateCount(resource_rep(qre.RZ, {"precision": 1e-9})),
-                ],
-            ),
-            (
-                2,
-                1e-3,
-                [
-                    GateCount(resource_rep(qre.RZ, {"precision": 1e-3}), 4),
-                    GateCount(resource_rep(qre.CNOT), 3),
-                ],
-            ),
-            (
-                5,
-                1e-5,
-                [
-                    GateCount(resource_rep(qre.RZ, {"precision": 1e-5}), (4**3) * 4),
-                    GateCount(resource_rep(qre.CNOT), (4**3) * 3),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Z",
-                                "num_ctrl_wires": 2,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        2 * 4**2,
-                    ),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Y",
-                                "num_ctrl_wires": 2,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        4**2,
-                    ),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Z",
-                                "num_ctrl_wires": 3,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        2 * 4,
-                    ),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Y",
-                                "num_ctrl_wires": 3,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        4,
-                    ),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Z",
-                                "num_ctrl_wires": 4,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        2,
-                    ),
-                    GateCount(
-                        resource_rep(
-                            qre.SelectPauliRot,
-                            {
-                                "rotation_axis": "Y",
-                                "num_ctrl_wires": 4,
-                                "precision": 1e-5,
-                            },
-                        ),
-                        1,
-                    ),
-                ],
-            ),
-        ),
-    )
-    def test_default_resources(self, num_wires, precision, expected_res):
-        """Test that the resources are correct."""
-        if precision is None:
-            config = ResourceConfig()
-            kwargs = config.resource_op_precisions[qre.QubitUnitary]
-            assert qre.QubitUnitary.resource_decomp(num_wires=num_wires, **kwargs) == expected_res
-        else:
-            assert (
-                qre.QubitUnitary.resource_decomp(num_wires=num_wires, precision=precision)
-                == expected_res
-            )
-
-
 class TestResourceSelectPauliRot:
     """Test the ResourceSelectPauliRot template"""
